Move RoutingPath enum in seperate file.
This resolve an issue with circular dependencies in the imports at runtime.
